refactor(core): unify stats data models and metrics computation across android and desktop
