+++ app/src/main/java/com/vinnovateit/latch/features/stats/SessionHistoryScreen.kt
@@ -39,6 +39,7 @@ import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.ui.semantics.Role
 import androidx.compose.ui.semantics.role
 import androidx.compose.ui.semantics.semantics
+import com.vinnovateit.latch.core.model.AggregatedDayRecord
 import com.vinnovateit.latch.features.stats.components.SessionHistorySkeletonLoader
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
@@ -55,6 +56,7 @@ import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.vinnovateit.latch.common.util.NoDataCard
 import com.vinnovateit.latch.common.util.formatBytes
+import com.vinnovateit.latch.core.model.DateRangeFilter
 import com.vinnovateit.latch.features.settings.manager.SettingsManager
 import com.vinnovateit.latch.features.stats.components.DayAggregateListItem
 import com.vinnovateit.latch.features.stats.components.GameStatRow
